refactor: 리뷰 반영
- api 파일명 camelCase로 변경
- cartAPI 추가
