+++ src/apis/PaymentAPI.ts
@@ -48,19 +48,3 @@ export const postNewAddressAPI = async ({
     throw error;
   }
 };
-// 봉달 목록 조회 API
-export const getCartsAPI = async () => {
-  try {
-    const res = await client.get('/carts');
-    return res.data;
-  } catch (error: any) {
-    if (error.response) {
-      // 서버 응답이 있는 경우 (오류 상태 코드 처리)
-      console.error('Server Error:', error.response.data);
-    } else {
-      // 서버 응답이 없는 경우 (네트워크 오류 등)
-      console.error('Error creating question:', error.message);
-    }
-    throw error;
-  }
-};
