+++ src/apis/index.ts
@@ -1,17 +1,14 @@
-import {
-  getAccessTokenAPI,
-  getOAuthLoginAPI,
-  getOAuthRedirectAPI,
-} from './AuthAPI';
-import {
-  getCartsAPI,
-  getDefaultAddressAPI,
-  postNewAddressAPI,
-} from './PaymentAPI';
 import { getAnnouncementsAPI, getBannersAPI } from './homeAPI';
 import { getProductsAPI, getProductDetailAPI } from './productAPI';
 import { getSearchProductsAPI, getTrendingKeywordsAPI } from './searchAPI';
 import { getReviewStatisticsAPI } from './reviewAPI';
+import {
+  getAccessTokenAPI,
+  getOAuthLoginAPI,
+  getOAuthRedirectAPI,
+} from './authAPI';
+import { getDefaultAddressAPI, postNewAddressAPI } from './paymentAPI';
+import { getCartsAPI } from './cartAPI';
 
 export {
   getBannersAPI,
